Move `setAndVerifyUserData` to `synctools` and replace with `AndroidAccountUtils`
