+++ core/src/main/kotlin/at/bitfire/davdroid/settings/migration/AccountSettingsMigration7.kt
@@ -9,7 +9,7 @@ import android.accounts.AccountManager
 import android.content.Context
 import android.provider.CalendarContract
 import at.bitfire.davdroid.settings.AccountSettings
-import at.bitfire.davdroid.sync.account.setAndVerifyUserData
+import at.bitfire.synctools.util.setAndVerifyUserData
 import at.bitfire.synctools.storage.calendar.AndroidCalendarProvider
 import dagger.Binds
 import dagger.Module
