+++ gradle/libs.versions.toml
@@ -33,6 +33,8 @@ kotlin = "2.3.21"
 kotlinx-coroutines = "1.11.0"
 ksp = "2.3.7"
 ktor = "3.4.3"
+# newer versions don't work with includeBuild("../synctools")
+# mikepenz-aboutLibraries = "14.0.1"
 mikepenz-aboutLibraries = "14.2.0"
 mockk = "1.14.9"
 okhttp = "5.3.2"
